api/requirements.txt
bidict==0.21.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SocketIO==5.0.3
gevent==21.1.2
gevent-websocket==0.10.1
google-auth==1.30.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
jenkinsapi==0.3.11
Jinja2==2.11.3
kubernetes==12.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-engineio==4.2.0
python-gitlab==2.7.1
python-socketio==5.3.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
